% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIP_stage.R, R/DIP_stage_impute.R
\name{DIP_stage}
\alias{DIP_stage}
\alias{DIP_stage_impute}
\title{Predicts the stage of host response dysregulation: Dysregulated Immune Profile 1-3}
\usage{
DIP_stage(new_data)

DIP_stage_impute(new_data)
}
\arguments{
\item{new_data}{A data frame containing the following columns:
\itemize{
\item \code{ID}: unique Identifier for the observations. Patients with multiple timepoints should have the timepoint included in their ID.
\item \code{TREM_1}: sTREM-1 measured in picograms per milliliter (pg/ml). Data should be untransformed and unscaled. Exact naming required. It is crucial that the
column name is exactly \code{TREM_1} as the function's code explicitly references this name.
\item \code{IL_6}: IL-6  measured in picograms per milliliter (pg/ml). Data should be untransformed and unscaled. Exact naming required. It is crucial that the
column name is exactly \code{IL_6} as the function's code explicitly references this name.
\item \code{Procalcitonin}: Procalcitonin  measured in picograms per milliliter (pg/ml). Data should be untransformed and unscaled. Exact naming required. It is crucial that the
column name is exactly \code{Procalcitonin} as the function's code explicitly references this name.
}}
}
\value{
A data frame with the original data alongside predictions and probability scores for each stage.

A data frame with the original data alongside predictions and probability scores for each stage.
}
\description{
This function uses an extreme XGBoost model to predict the degree of host response dysregulation using plasma concentrations of sTREM-1, IL-6, and Procalcitonin.
The prediction model leverages absolute values to deliver precise, tailored outcomes, suitable for single-patient scenarios
The function automatically omits patients with missing classifier data.
Outputs include the original classifier data, the predicted DIP stage, and probability scores for each predicted stage.
DIP1 represent minor, DIP2 moderate and DIP3 major host response dysregulation.
Interactive visualizations are also provided when run in an appropriate environment.

This function uses an extreme XGBoost model to predict the degree of host response dysregulation using using plasma concentrations of sTREM-1, IL-6, and Procalcitonin.
The prediction model leverages absolute values to deliver precise, tailored outcomes, suitable for single-patient scenarios
The function imputes missing values using 10 generated MICE datasets if necessary.
Be aware that this 3-biomarker based model with one imputed classifier has not been validated.
Outputs include the original classifier data, the predicted DIP stage, and probability scores for each predicted stage.
DIP1 represent minor, DIP2 moderate and DIP3 major host response dysregulation.
Interactive visualizations are also provided when run in an appropriate environment.
}
\examples{
test_data <- data.frame(
  ID = 1:3,
  TREM_1 = c(182, 400, 1000),
  IL_6 = c(70, 5, 10000),
  Procalcitonin = c(877, 66, 20000)
)
result <- DIP_stage(test_data)
print(result)
test_data <- data.frame(
  ID = 1:3,
  TREM_1 = c(182, 400, 1000),
  IL_6 = c(70, 5, 10000),
  Procalcitonin = c(877, 66, 20000)
)
result <- DIP_stage(test_data)
print(result)
}
