% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny wappers.R
\name{run_DIP_stage_for_shiny}
\alias{run_DIP_stage_for_shiny}
\title{Shiny-compatible wrapper for DIP_stage()}
\usage{
run_DIP_stage_for_shiny(new_data)
}
\description{
Shiny-compatible wrapper for DIP_stage()
}
