% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cDIP.R
\name{cDIP}
\alias{cDIP}
\title{Predicts the continuous immune dysregulation score (cDIP)}
\usage{
cDIP(new_data)
}
\arguments{
\item{new_data}{A data frame containing the following columns:
\itemize{
\item \code{ID}: unique identifier for the observations. Patients with multiple timepoints should have the timepoint included in their ID.
\item \code{TREM_1}: sTREM-1 measured in picograms per milliliter (pg/ml). Data should be untransformed and unscaled. The column name must be exactly \code{TREM_1}.
\item \code{IL_6}: IL-6 measured in picograms per milliliter (pg/ml). Data should be untransformed and unscaled. The column name must be exactly \code{IL_6}.
\item \code{Procalcitonin}: Procalcitonin measured in picograms per milliliter (pg/ml). Data should be untransformed and unscaled. The column name must be exactly \code{Procalcitonin}.
}}
}
\value{
A data frame with the predicted continuous immune dysregulation score (cDIP):
\itemize{
\item \code{ID}: Unique identifier for each observation.
\item \code{cDIP}: Continuous immune dysregulation score (0-1).
}
}
\description{
This function uses a machine-learning model (random forest regressor) to predict the continuous degree of immune dysregulation (cDIP) using plasma concentrations of sTREM-1, IL-6, and Procalcitonin.
The model leverages absolute values to deliver precise, tailored outcomes, suitable for single-patient scenarios.
The function automatically omits patients with missing classifier data.
Outputs include the original classifier data and the predicted continuous dysregulation score (cDIP), ranging from 0 to 1.
A higher cDIP score reflects a greater degree of immune dysregulation.
An interactive beeswarm plot is generated to visualize the distribution of cDIP scores.
The prediction results and the plot are saved to the global environment.
}
\examples{
test_data <- data.frame(
  ID = 1:20,
  TREM_1 = c(182, 400, 1000, 560, 230, 900, 450, 710, 620, 350,
             150, 800, 250, 490, 780, 340, 900, 1100, 220, 510),
  IL_6 = c(70, 5, 10000, 450, 88, 3000, 150, 680, 740, 50,
           30, 600, 120, 470, 800, 60, 5000, 9000, 33, 200),
  Procalcitonin = c(877, 66, 20000, 1500, 500, 10000, 800, 2700, 1800, 460,
                    250, 12000, 600, 1100, 14000, 350, 15000, 18000, 310, 900)
)
cDIP(test_data)
}
