% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIP_stage.R
\name{DIP_stage}
\alias{DIP_stage}
\title{Predicts the stage of host response dysregulation: Dysregulated Immune Profile 1-3}
\usage{
DIP_stage(new_data)
}
\arguments{
\item{new_data}{A data frame containing the following columns:
\itemize{
\item \code{ID}: unique identifier for the observations. Patients with multiple timepoints should have the timepoint included in their ID.
\item \code{TREM_1}: sTREM-1 measured in picograms per milliliter (pg/ml). Data should be untransformed and unscaled. The column name must be exactly \code{TREM_1}.
\item \code{IL_6}: IL-6 measured in picograms per milliliter (pg/ml). Data should be untransformed and unscaled. The column name must be exactly \code{IL_6}.
\item \code{Procalcitonin}: Procalcitonin measured in picograms per milliliter (pg/ml). Data should be untransformed and unscaled. The column name must be exactly \code{Procalcitonin}.
}}
}
\value{
A data frame with predicted DIP stage and probabilities:
#' \itemize{
\item \code{ID}: Unique identifier for each observation.
\item \code{DIP}: Predicted DIP stage (DIP1, DIP2, or DIP3).
\item \code{DIP1_Prob}, \code{DIP2_Prob}, \code{DIP3_Prob}: Prediction probabilities for each stage.
}
}
\description{
This function uses an extreme XGBoost model to predict the degree of host response dysregulation using plasma concentrations of sTREM-1, IL-6, and Procalcitonin.
The prediction model leverages absolute values to deliver precise, tailored outcomes, suitable for single-patient scenarios.
The function automatically omits patients with missing classifier data.
Outputs include the original classifier data, the predicted DIP stage, and probability scores for each predicted stage.
DIP1 represents minor, DIP2 moderate and DIP3 major host response dysregulation.
A pie chart (using plotly) is generated for the distribution of predictions, and an interactive 3D scatter plot is also created.
The prediction results and both plots are saved to the global environment.
}
\examples{
test_data <- data.frame(
  ID = 1:3,
  TREM_1 = c(182, 400, 1000),
  IL_6 = c(70, 5, 10000),
  Procalcitonin = c(877, 66, 20000)
)
DIP_stage(test_data)
}
